backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
pydantic
google-cloud-storage
python-dotenv

